유저 수정, 삭제 API 예외처리

+++ 개발 기초/library-app/library-app/src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -42,12 +42,24 @@ public class UserController {
 
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request){
+        String readsql = "SELECT * FROM user WHERE id = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readsql, (rs, rowNum) -> 0,request.getId()).isEmpty();
+        if(isUserNotExist){
+            throw new IllegalArgumentException();
+        }
+
         String sql = "UPDATE user SET name = ? WHERE id = ?";
         jdbcTemplate.update(sql, request.getName(), request.getId());
     }
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name){
+        String readsql = "SELECT * FROM user WHERE name = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readsql, (rs, rowNum) -> 0,name).isEmpty();
+        if(isUserNotExist){
+            throw new IllegalArgumentException();
+        }
+
         String sql = "DELETE FROM user WHERE name = ?";
         jdbcTemplate.update(sql, name);
     }
